requirements.txt
#flask==2.1.1
flask==3.1.0

waitress==2.1.2
#openai==0.10.2
#openai==1.57.0
openai==0.28.0

python-dotenv==0.21.0
flask-Cors==3.0.10
fpdf==1.7.2
